test (#76): Add specs for custom field props
Now that we're allowing users to set props for custom fields
(e.g. setting a custom field to determine the label of an option,
or a custom field to indicate which property of an option is guaranteed to be its unique identifier),
we need to test that those props actually work:
when a fieldLabel prop is set, will the options still display in the HTML correctly?
and when a uniqueIdLabel prop is set, will the options get selected correctly & set our v-model values correctly?