test (#78): Add SelectSingle compare-value specs
Like the specs added to SelectMulti, these specs will check whether SelectSingle
is able to compare the `options` array to the list of currently selected options using **value**,
rather than **reference**

Additional work: refactor SelectSingle specs to match extracted-shared-code approach of SelectMulti approach
TODO: 2 specs for SelectSingle are currently failing, and this seems to indicate a possible bug with the handling of `uniqueIdField` in SelectSingle -
this needs to be fixed. Yay specs!